move reference variables into own folder and update imports
