--- src/user.ts
@@ -1,2 +1,0 @@
-export const emailJsUser = "user_ibvDbNtHebP8cT2vHy7ut";
-export const emailJsService = "service_huafz7u";
