--- src/utils/greet.ts
@@ -1,3 +1,0 @@
-export function greet(name: string): string {
-  return `Hello, ${name}!`;
-}
